+++ mcumgr-toolkit-cli/src/args.rs
@@ -40,7 +40,7 @@ pub struct App {
     pub baud: u32,
 
     /// Communication timeout (in ms)
-    #[arg(short, long, default_value_t = 10000)]
+    #[arg(short, long, default_value_t = 500)]
     pub timeout: u64,
 
     /// Settings that customize runtime behaviour