+++ mcumgr-toolkit/src/client.rs
@@ -38,6 +38,7 @@ const ZEPHYR_DEFAULT_SMP_FRAME_SIZE: usize = 384;
 pub struct MCUmgrClient {
     connection: Connection,
     smp_frame_size: AtomicUsize,
+    num_retries: u8,
 }
 
 /// Possible error values of [`MCUmgrClient`].
@@ -220,7 +221,7 @@ impl MCUmgrClient {
     /// # use mcumgr_toolkit::MCUmgrClient;
     /// # fn main() {
     /// let serial = serialport::new("COM42", 115200)
-    ///     .timeout(std::time::Duration::from_millis(10000))
+    ///     .timeout(std::time::Duration::from_millis(500))
     ///     .open()
     ///     .unwrap();
     ///
@@ -233,6 +234,7 @@ impl MCUmgrClient {
         Self {
             connection: Connection::new(SerialTransport::new(serial)),
             smp_frame_size: ZEPHYR_DEFAULT_SMP_FRAME_SIZE.into(),
+            num_retries: 5, // Hard coded for now
         }
     }
 
@@ -349,7 +351,7 @@ impl MCUmgrClient {
     pub fn use_auto_frame_size(&self) -> Result<(), MCUmgrClientError> {
         let mcumgr_params = self
             .connection
-            .execute_command(&commands::os::MCUmgrParameters)?;
+            .execute_command(&commands::os::MCUmgrParameters, self.num_retries)?;
 
         log::debug!("Using frame size {}.", mcumgr_params.buf_size);
 
@@ -415,7 +417,7 @@ impl MCUmgrClient {
     /// This can be used as a sanity check for whether the device is connected and responsive.
     pub fn os_echo(&self, msg: impl AsRef<str>) -> Result<String, MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::os::Echo { d: msg.as_ref() })
+            .execute_command(&commands::os::Echo { d: msg.as_ref() }, self.num_retries)
             .map(|resp| resp.r)
             .map_err(Into::into)
     }
@@ -434,7 +436,7 @@ impl MCUmgrClient {
         &self,
     ) -> Result<HashMap<String, commands::os::TaskStatisticsEntry>, MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::os::TaskStatistics)
+            .execute_command(&commands::os::TaskStatistics, self.num_retries)
             .map(|resp| {
                 let mut tasks = resp.tasks;
                 for (_, stats) in tasks.iter_mut() {
@@ -452,7 +454,7 @@ impl MCUmgrClient {
         datetime: chrono::NaiveDateTime,
     ) -> Result<(), MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::os::DateTimeSet { datetime })
+            .execute_command(&commands::os::DateTimeSet { datetime }, self.num_retries)
             .map(Into::into)
             .map_err(Into::into)
     }
@@ -460,7 +462,7 @@ impl MCUmgrClient {
     /// Retrieves the device RTC's datetime.
     pub fn os_get_datetime(&self) -> Result<chrono::NaiveDateTime, MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::os::DateTimeGet)
+            .execute_command(&commands::os::DateTimeGet, self.num_retries)
             .map(|val| val.datetime)
             .map_err(Into::into)
     }
@@ -484,7 +486,10 @@ impl MCUmgrClient {
         boot_mode: Option<u8>,
     ) -> Result<(), MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::os::SystemReset { force, boot_mode })
+            .execute_command(
+                &commands::os::SystemReset { force, boot_mode },
+                self.num_retries,
+            )
             .map(Into::into)
             .map_err(Into::into)
     }
@@ -511,7 +516,7 @@ impl MCUmgrClient {
     ///
     pub fn os_application_info(&self, format: Option<&str>) -> Result<String, MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::os::ApplicationInfo { format })
+            .execute_command(&commands::os::ApplicationInfo { format }, self.num_retries)
             .map(|resp| resp.output)
             .map_err(Into::into)
     }
@@ -521,14 +526,15 @@ impl MCUmgrClient {
         Ok(
             match self
                 .connection
-                .execute_command(&commands::os::BootloaderInfo)?
+                .execute_command(&commands::os::BootloaderInfo, self.num_retries)?
                 .bootloader
                 .as_str()
             {
                 "MCUboot" => {
-                    let mode_data = self
-                        .connection
-                        .execute_command(&commands::os::BootloaderInfoMcubootMode {})?;
+                    let mode_data = self.connection.execute_command(
+                        &commands::os::BootloaderInfoMcubootMode {},
+                        self.num_retries,
+                    )?;
                     BootloaderInfo::MCUboot {
                         mode: mode_data.mode,
                         no_downgrade: mode_data.no_downgrade,
@@ -544,7 +550,7 @@ impl MCUmgrClient {
     /// Obtain a list of images with their current state.
     pub fn image_get_state(&self) -> Result<Vec<commands::image::ImageState>, MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::image::GetImageState)
+            .execute_command(&commands::image::GetImageState, self.num_retries)
             .map(|val| val.images)
             .map_err(Into::into)
     }
@@ -570,10 +576,13 @@ impl MCUmgrClient {
         confirm: bool,
     ) -> Result<Vec<commands::image::ImageState>, MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::image::SetImageState {
-                hash: hash.as_ref(),
-                confirm,
-            })
+            .execute_command(
+                &commands::image::SetImageState {
+                    hash: hash.as_ref(),
+                    confirm,
+                },
+                self.num_retries,
+            )
             .map(|val| val.images)
             .map_err(Into::into)
     }
@@ -621,25 +630,29 @@ impl MCUmgrClient {
             let chunk_data = &data[offset..offset + current_chunk_size];
 
             let upload_response = if offset == 0 {
-                self.connection
-                    .execute_command(&commands::image::ImageUpload {
+                self.connection.execute_command(
+                    &commands::image::ImageUpload {
                         image,
                         len: Some(size as u64),
                         off: offset as u64,
                         sha: Some(&actual_checksum),
                         data: chunk_data,
                         upgrade: Some(upgrade_only),
-                    })?
+                    },
+                    self.num_retries,
+                )?
             } else {
-                self.connection
-                    .execute_command(&commands::image::ImageUpload {
+                self.connection.execute_command(
+                    &commands::image::ImageUpload {
                         image: None,
                         len: None,
                         off: offset as u64,
                         sha: None,
                         data: chunk_data,
                         upgrade: None,
-                    })?
+                    },
+                    self.num_retries,
+                )?
             };
 
             offset = upload_response
@@ -681,7 +694,7 @@ impl MCUmgrClient {
     ///
     pub fn image_erase(&self, slot: Option<u32>) -> Result<(), MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::image::ImageErase { slot })
+            .execute_command(&commands::image::ImageErase { slot }, self.num_retries)
             .map(Into::into)
             .map_err(Into::into)
     }
@@ -691,7 +704,7 @@ impl MCUmgrClient {
         &self,
     ) -> Result<Vec<commands::image::SlotInfoImage>, MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::image::SlotInfo)
+            .execute_command(&commands::image::SlotInfo, self.num_retries)
             .map(|val| val.images)
             .map_err(Into::into)
     }
@@ -716,9 +729,10 @@ impl MCUmgrClient {
         mut progress: Option<&mut dyn FnMut(u64, u64) -> bool>,
     ) -> Result<(), MCUmgrClientError> {
         let name = name.as_ref();
-        let response = self
-            .connection
-            .execute_command(&commands::fs::FileDownload { name, off: 0 })?;
+        let response = self.connection.execute_command(
+            &commands::fs::FileDownload { name, off: 0 },
+            self.num_retries,
+        )?;
 
         let file_len = response.len.ok_or(MCUmgrClientError::MissingSize)?;
         if response.off != 0 {
@@ -745,9 +759,10 @@ impl MCUmgrClient {
         }
 
         while offset < file_len {
-            let response = self
-                .connection
-                .execute_command(&commands::fs::FileDownload { name, off: offset })?;
+            let response = self.connection.execute_command(
+                &commands::fs::FileDownload { name, off: offset },
+                self.num_retries,
+            )?;
 
             if response.off != offset {
                 return Err(MCUmgrClientError::UnexpectedOffset);
@@ -814,12 +829,15 @@ impl MCUmgrClient {
                 .read_exact(chunk_buffer)
                 .map_err(MCUmgrClientError::ReaderError)?;
 
-            self.connection.execute_command(&commands::fs::FileUpload {
-                off: offset,
-                data: chunk_buffer,
-                name,
-                len: if offset == 0 { Some(size) } else { None },
-            })?;
+            self.connection.execute_command(
+                &commands::fs::FileUpload {
+                    off: offset,
+                    data: chunk_buffer,
+                    name,
+                    len: if offset == 0 { Some(size) } else { None },
+                },
+                self.num_retries,
+            )?;
 
             offset += chunk_buffer.len() as u64;
 
@@ -878,7 +896,7 @@ impl MCUmgrClient {
         &self,
     ) -> Result<HashMap<String, commands::fs::FileChecksumProperties>, MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::fs::SupportedFileChecksumTypes)
+            .execute_command(&commands::fs::SupportedFileChecksumTypes, self.num_retries)
             .map(|val| val.types)
             .map_err(Into::into)
     }
@@ -886,7 +904,7 @@ impl MCUmgrClient {
     /// Close all device files MCUmgr has currently open
     pub fn fs_file_close(&self) -> Result<(), MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::fs::FileClose)
+            .execute_command(&commands::fs::FileClose, self.num_retries)
             .map(Into::into)
             .map_err(Into::into)
     }
@@ -902,7 +920,7 @@ impl MCUmgrClient {
     /// A tuple of (returncode, stdout) produced by the command execution.
     pub fn shell_execute(&self, argv: &[String]) -> Result<(i32, String), MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::shell::ShellCommandLineExecute { argv })
+            .execute_command(&commands::shell::ShellCommandLineExecute { argv }, 0)
             .map(|ret| (ret.ret, ret.o))
             .map_err(Into::into)
     }
@@ -910,7 +928,7 @@ impl MCUmgrClient {
     /// Erase the `storage_partition` flash partition.
     pub fn zephyr_erase_storage(&self) -> Result<(), MCUmgrClientError> {
         self.connection
-            .execute_command(&commands::zephyr::EraseStorage)
+            .execute_command(&commands::zephyr::EraseStorage, self.num_retries)
             .map(Into::into)
             .map_err(Into::into)
     }